pyproject.toml
[project]
name = "nonebot_plugin_suggarex_cf"
version = "1.1.5"
description = "适用于SuggarChat插件的CloudFlare协议适配器实现，SuggarChat附属插件"
authors = [{ name = "JohnRichard4096", email = "windowserror@163.com" }]
dependencies = [
    "nonebot2>=2.2.0",
    "nonebot-plugin-suggarchat>=3.1.0",
    "nonebot-adapter-onebot>=2.0.0",
    "aiohttp>=3.9.0",
]
requires-python = ">=3.10, <4.0"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
keywords = ['chat", "suggar", "nonebot']

[project.urls]
"Homepage" = "https://github.com/LiteSuggarDEV/nonebot_plugin_suggarex_cf"
"Source" = "https://github.com/LiteSuggarDEV/nonebot_plugin_suggarex_cf"
"Issue Tracker" = "https://github.com/LiteSuggarDEV/nonebot_plugin_suggarex_cf/issues"
[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
    "I",     # isort
    "PERF",  # pylint-performance
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "E501",   # line-too-long
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.pyright]
typeCheckingMode = "standard"